Upgrade to statinamic 0.7.0
