+++ .gitignore
@@ -9,4 +9,3 @@ design/node_modules
 
 # statinamic builded files
 dist
-scripts/statinamic-static.js
