+++ package.json
@@ -9,22 +9,43 @@
     "eslint-staged": "eslint-staged",
     "stylelint-staged": "stylelint-staged",
     "svgo-staged": "./scripts/svgo.sh",
-    "start": "babel-node scripts/build --server --dev --open",
-    "static": "babel-node scripts/build --static --production",
+    "statinamic": "cross-env BABEL_ENV=statinamic DEBUG=statinamic:* babel-node scripts/build",
+    "start": "npm run statinamic -- --server --open --dev",
+    "build": "npm run statinamic -- --static --production",
     "pretest": "npm run lint",
-    "test": "npm run static",
-    "posttest": "echo 'postcss.org' > dist/CNAME && touch dist/.nojekyll",
+    "test": "npm run build",
     "styleguide-server": "styleguidist server",
     "styleguide-build": "styleguidist build",
     "predeploy": "npm run build",
     "deploy": "./scripts/deploy.sh"
   },
+  "statinamic": {
+    "assets": false,
+    "CNAME": true
+  },
   "babel": {
-    "stage": 0
+    "presets": [
+      "babel-preset-react",
+      "babel-preset-es2015",
+      "babel-preset-stage-1"
+    ],
+    "env": {
+      "statinamic": {
+        "plugins": [
+          [
+            "babel-plugin-webpack-loaders",
+            {
+              "config": "./scripts/webpack.config.babel.js",
+              "verbose": false
+            }
+          ]
+        ]
+      }
+    }
   },
   "eslintConfig": {
+    "parser": "babel-eslint",
     "extends": [
-      "eslint-config-i-am-meticulous",
       "eslint-config-i-am-meticulous/react"
     ]
   },
@@ -37,21 +58,25 @@
     "svgo-staged"
   ],
   "devDependencies": {
-    "babel": "^5.8.34",
-    "babel-core": "^5.8.34",
+    "babel-cli": "^6.3.17",
+    "babel-core": "^6.3.26",
     "babel-eslint": "^4.1.6",
-    "babel-loader": "^5.4.0",
-    "babel-plugin-react-transform": "^1.1.1",
+    "babel-loader": "^6.2.0",
+    "babel-plugin-webpack-loaders": "^0.2.2",
+    "babel-preset-es2015": "^6.3.13",
+    "babel-preset-react": "^6.3.13",
+    "babel-preset-react-hmre": "^1.0.1",
+    "babel-preset-stage-1": "^6.3.13",
+    "cross-env": "^1.0.7",
     "css-loader": "^0.17.0",
     "eslint": "^1.10.3",
-    "eslint-config-i-am-meticulous": "^1.0.0",
+    "eslint-config-i-am-meticulous": "^2.0.0",
     "eslint-loader": "^1.2.0",
     "eslint-plugin-react": "^3.14.0",
     "extract-text-webpack-plugin": "^0.8.2",
     "file-loader": "^0.8.5",
     "fontfaceobserver": "^1.6.3",
     "highlight.js": "^8.9.1",
-    "history": "^1.17.0",
     "invariant": "^2.2.0",
     "json-loader": "^0.5.4",
     "lint-staged": "^0.1.1",
@@ -60,30 +85,27 @@
     "postcss-cssnext": "^2.4.0",
     "postcss-loader": "^0.7.0",
     "pre-commit": "^1.1.2",
-    "react": "^0.14.6",
-    "react-dom": "^0.14.6",
+    "react": "^0.14.7",
+    "react-dom": "^0.14.7",
     "react-google-analytics": "^0.2.0",
     "react-helmet": "^2.2.0",
-    "react-redux": "^2.1.2",
-    "react-router": "^1.0.3",
+    "react-redux": "^4.0.6",
+    "react-router": "^2.0.0",
     "react-styleguidist": "^1.3.2",
-    "react-transform-catch-errors": "^1.0.1",
-    "react-transform-hmr": "^1.0.1",
-    "redbox-react": "^1.2.0",
-    "redux": "^2.0.0",
-    "redux-devtools": "^2.1.5",
-    "redux-thunk": "^0.1.0",
-    "staged-files": "^0.1.2",
-    "statinamic": "^0.5.1",
-    "style-loader": "^0.12.4",
-    "stylelint": "^3.2.2",
-    "stylelint-config-standard": "^0.2.0",
     "svgo": "^0.6.1",
+    "redux": "^3.0.0",
+    "redux-devtools": "^3.0.1",
+    "redux-devtools-dock-monitor": "^1.0.1",
+    "redux-devtools-log-monitor": "^1.0.1",
+    "redux-thunk": "^1.0.3",
+    "statinamic": "^0.7.0",
+    "style-loader": "^0.12.4",
+    "stylelint": "^4.0.0",
+    "stylelint-config-standard": "^2.0.0",
     "webpack": "^1.12.9",
     "whatwg-fetch": "^0.9.0"
   },
   "dependencies": {
-    "highlight.js": "^8.9.1",
     "react-highlight": "^0.6.1"
   }
 }
