+++ scripts/config.js
@@ -1,0 +1,6 @@
+import pkg from "../package.json"
+import configurator from "statinamic/lib/configurator"
+
+const config = configurator(pkg)
+
+export default config
