+++ web_modules/LayoutContainer/index.js
@@ -2,10 +2,10 @@ import React, { Component } from "react"
 import { PropTypes } from "react"
 import Helmet from "react-helmet"
 
-import Navigation from "Navigation"
-import Social from "Social"
-import Footer from "Footer"
-import GoogleAnalyticsTracker from "GoogleAnalyticsTracker"
+import Navigation from "../Navigation"
+import Social from "../Social"
+import Footer from "../Footer"
+import GoogleAnalyticsTracker from "../GoogleAnalyticsTracker"
 
 import styles from "./index.css"
 
@@ -30,7 +30,7 @@ if (typeof window !== "undefined") {
   })
 }
 
-export default class Wrapper extends Component {
+export default class LayoutContainer extends Component {
 
   static propTypes = {
     children: PropTypes.oneOfType([ PropTypes.array, PropTypes.object ]),
