+++ web_modules/app/metadata.js
@@ -1,0 +1,7 @@
+import pkg from "../../package.json"
+
+export default {
+  pkg,
+
+  // you can add any data you want here
+}
