+++ web_modules/layouts/index.js
@@ -1,0 +1,3 @@
+export Page from "./Page" // default layout
+export PageError from "./PageError"
+export HomePage from "./HomePage"
